refactor: moved getPosition in a dedicated file.
