+++ src/utils/getPosition.ts
@@ -1,0 +1,12 @@
+import * as vscode from "vscode";
+
+export function getPosition(document: vscode.TextDocument, packageName: string) {
+  const regex = new RegExp(`"${packageName}"\\s*:`);
+  const line = document
+    .getText()
+    .split("\n")
+    .findIndex((line) => regex.test(line));
+  const character = document.lineAt(line).text.indexOf(`"${packageName}":`);
+
+  return { line, character };
+}
